Updated HTML to link to vid.html and pic.html for video and photo choices
This commit updates the choice.html file by changing the href attribute in the first and second anchor tags. The first anchor tag now links to vid.html, and the second anchor tag now links to pic.html. These changes were made to allow the user to choose between watching a video or viewing a photo. The alert functions in the onclick attributes remain unchanged.