Add additional Boomi images and add carousel images
This diff adds 9 new Boomi images (6.jpg to 14.jpg) and 7 carousel images (1.jpg to 7.jpg) to the HTML file. The new Boomi images are placed after the existing ones and the carousel images are added after the Boomi images. The carousel images are also added in reverse order at the end of the file.