Change hardcoded password to match username in index.html login
This commit updates the hardcoded password to match the username in the index.html file for the login form. The if condition in the JavaScript code now checks if the password matches the username, i.e., "boomica19@birthday.com". Previously, the password was set to a number (27505), which has been changed to match the username for consistency.